+++ libs/designer/src/lib/core/state/selectors/actionMetadataSelector.ts
@@ -34,13 +34,13 @@ export const useNodeConnectionName = (nodeId: string): QueryResult => {
     () =>
       nodeId && connectionId
         ? {
-          isLoading,
-          result: !isLoading ? connection?.properties?.displayName ?? connectionId.split('/').at(-1) : '',
-        }
+            isLoading,
+            result: !isLoading ? connection?.properties?.displayName ?? connectionId.split('/').at(-1) : '',
+          }
         : {
-          isLoading: false,
-          result: undefined,
-        },
+            isLoading: false,
+            result: undefined,
+          },
     [nodeId, connection?.properties?.displayName, connectionId, isLoading]
   );
 };
@@ -63,7 +63,10 @@ export const useOutputParameters = (nodeId: string) => {
   return useSelector((state: RootState) => selector(state, nodeId));
 };
 
-export const useOperationManifest = (operationInfo?: NodeOperation, enabled = true): UseQueryResult<OperationManifest | undefined, unknown> => {
+export const useOperationManifest = (
+  operationInfo?: NodeOperation,
+  enabled = true
+): UseQueryResult<OperationManifest | undefined, unknown> => {
   const operationManifestService = OperationManifestService();
   const connectorId = operationInfo?.connectorId?.toLowerCase();
   const operationId = operationInfo?.operationId?.toLowerCase();
@@ -144,6 +147,46 @@ export const useOperationSummary = (operationInfo: NodeOperation): QueryResult =
   const result = useNodeAttributeOrSwagger(operationInfo, ['summary'], ['summary'], 'summary', { useManifest });
   if (result.result === undefined && operationInfo?.operationId) {
     result.result = titleCase(operationInfo.operationId);
+  }
+
+  return result;
+};
+
+export const useOperationUploadChunkMetadata = (operationInfo: NodeOperation): QueryResult => {
+  const operationManifestService = OperationManifestService();
+  const useManifest = operationManifestService.isSupported(operationInfo?.type ?? '', operationInfo?.kind ?? '');
+
+  const result = useNodeAttributeOrSwagger(operationInfo, ['settings', 'chunking'], ['uploadChunkMetadata'], 'uploadChunkMetadata', {
+    useManifest,
+  });
+
+  if (!result.isLoading) {
+    if (useManifest && result?.result) {
+      return { result: result.result.options, isLoading: false };
+    }
+  }
+
+  return result;
+};
+
+export const useOperationDownloadChunkMetadata = (operationInfo: NodeOperation): QueryResult => {
+  const operationManifestService = OperationManifestService();
+  const useManifest = operationManifestService.isSupported(operationInfo?.type ?? '', operationInfo?.kind ?? '');
+
+  const result = useNodeAttributeOrSwagger(
+    operationInfo,
+    ['settings', 'downloadChunking'],
+    ['downloadChunkMetadata'],
+    'downloadChunkMetadata',
+    {
+      useManifest,
+    }
+  );
+
+  if (!result.isLoading) {
+    if (useManifest && result.result) {
+      return { result: result.result.options, isLoading: false };
+    }
   }
 
   return result;